Mostly-fix Trac #2595: updates for existentials
Ganesh wanted to update records that involve existentials.  That 
seems reasonable to me, and this patch covers existentials, GADTs,
and data type families.

The restriction is that 
  The types of the updated fields may mention only the
  universally-quantified type variables of the data constructor

This doesn't allow everything in #2595 (it allows 'g' but not 'f' in
the ticket), but it gets a lot closer.

Lots of the new lines are comments!