Fix Trac #2720: inlining and casts
The issue here is what happens when we have

	(f |> co) x

where f is itself marked INLINE.  We want callSiteInline to "see" 
the fact that the function is applied, and hence have some incentive
to inline.  I've done this by extending CoreUnfold.CallCtxt with 
ValAppCtxt.  I think that should catch this case without messing up
any of the others.