Avoid nasty name clash with associated data types (fixes Trac #2888)
The main bug was in TcHsType; see Note [Avoid name clashes for 
associated data types].  However I did a bit of re-factoring while 
I was abouut it.

I'm still a but unhappy with the use of TyCon.setTyConArgPoss; it'd
be better to construct the TyCon correctly in the first place.  But
that means passing an extra parameter to tcTyDecl1... maybe we should
do this.