soup-up "ghc-pkg check"
We now look for missing files (including .hi files), and report all
the packages that are transitively broken.

$ ghc-pkg check 
There are problems in package syb-0.1.0.0:
  dependency foo-4.0.0.0 doesn't exist
There are problems in package process-1.0.1.1:
  file System/Process.hi is missing

The following packages are broken, either because they have a problem
listed above, or because they depend on a broken package.
syb-0.1.0.0
process-1.0.1.1
base-3.0.3.0
Cabal-1.7.0
haskell98-1.0.1.0
haddock-2.4.2
ghc-6.11