Improve transferPolyIdInfo for value-arg abstraction
If we float a binding out of a *value* lambda, the fixing-up of IdInfo
is a bit more complicated than before.  Since in principle FloatOut
can do this (and thus can do full lambda lifting), it's imporrtant
that transferPolyIdInfo does the Right Thing.

This doensn't matter unless you use FloatOut's abilty to lambda-lift, 
which GHC mostly doesn't, yet.  But Max used it and tripped over this bug.