Fix Trac #246: order of matching in record patterns
While I was looking at the desugaring of pattern matching (fixing
Trac #3126) I finally got around to fixing another long-standing bug:
when matching in a record pattern, GHC should match left-to-right in
the programmer-specfied order, *not* left-to-right positionally in 
the original record declaration.

Needless to say, that requires a little more code. 
See Note [Record patterns] in MatchCon.lhs