Worker/wrapper should make INLINE if it doesn't w/w
If worker/wrapper decides not to w/w something on the grounds that
it's too small, it should add an INLINE pragma.  Otherwise, later
in the day that small thing might now be big, and we'd wish we'd
done the w/w after all.  This only made a difference in one nofib 
program (bspt), but it's an easy change.

See Note [Don't w/w inline things (a) and (b)]