Improve printing of Orig RdrNames
In Tempate Haskell -ddump-splices, the "after" expression is populated 
with RdrNames, many of which are Orig things.  We used to print these
fully-qualified, but that's a bit heavy.

This patch refactors the code a bit so that the same print-unqualified
mechanism we use for Names also works for RdrNames.  Lots of comments
too, because it took me a while to figure out how it all worked again.