Make two type defaults explicit
Now that -Werror rejects programs that use silent type-class defaulting,
we must commit in the source code.

I've used Double in CPUTime, which is the same as was picked automatically
before, but I expect Float would be ok.

   realToInteger :: Real a => a -> Integer
   realToInteger ct = round (realToFrac ct :: Double)

In GHC.Float I used Float (rather that than the auto-picked Double)
because I'm pretty certain it has enough precision.

	-- f :: Integer, log :: Float -> Float, 
        --               ceiling :: Float -> Int
        ceiling ((log (fromInteger (f+1) :: Float) +