Refactor, and improve error messages (cf Trac #3395)
The Convert stuff should not panic if the programmer hands over an
invalid TH term; instead it should give a graceful error message.
Largely this had been done, but not for do-blocks, so this patch
fixes that problem.

Moreover, I did some refactoring and tidying up, which is why
so many lines of code have changed